refactor default strips
