+++ libs/light/defaultlights.ts
@@ -5,64 +5,6 @@ namespace light {
     //% whenUsed
     export const pixels = light.defaultStrip();
 
-    let _defaultStrip: NeoPixelStrip;
-    /**
-     * Get the default light strip.
-     */
-    //% help=light/default-strip
-    //% blockId="neopixel_default_strip" block="default strip"
-    //% weight=110 blockGap=8
-    //% advanced=true
-    export function defaultStrip(): NeoPixelStrip {
-        if (_defaultStrip) return _defaultStrip;
-
-        const data = pins.pinByCfg(DAL.CFG_PIN_DOTSTAR_DATA);
-        const clk = pins.pinByCfg(DAL.CFG_PIN_DOTSTAR_CLOCK);
-        const dsnum = control.getConfigValue(DAL.CFG_NUM_DOTSTARS, 0);
-        const neo = pins.pinByCfg(DAL.CFG_PIN_NEOPIXEL);
-        const neonum = control.getConfigValue(DAL.CFG_NUM_DOTSTARS, 0);
-        const mosi = pins.pinByCfg(DAL.CFG_PIN_MOSI);
-        const sck = pins.pinByCfg(DAL.CFG_PIN_SCK);
-
-        if (data && clk && dsnum > 1) {
-            _defaultStrip = light.createAPA102Strip(data, clk, dsnum);
-            _defaultStrip.setBrightness(96);
-        } else if(neo && neonum > 1) {
-            _defaultStrip = light.createNeoPixelStrip(neo, neonum, NeoPixelMode.RGB);
-        } else { // mount strip on SPI
-            _defaultStrip = light.createAPA102Strip(mosi, sck, 30);
-        }
-
-        return _defaultStrip;
-    }
-
-    let _onboardStrip: light.LightStrip;
-    /**
-     * Get the default light strip.
-     */
-    //% help=light/onboard-strip
-    //% blockId="neopixel_onboard_strip" block="onboard strip"
-    //% weight=111 blockGap=8
-    //% advanced=true
-    export function onboardStrip(): NeoPixelStrip {
-        if (_onboardStrip) return _onboardStrip;
-
-        const data = pins.pinByCfg(DAL.CFG_PIN_DOTSTAR_DATA);
-        const clk = pins.pinByCfg(DAL.CFG_PIN_DOTSTAR_CLOCK);
-        const dsnum = control.getConfigValue(DAL.CFG_NUM_DOTSTARS, 1);
-        const neo = pins.pinByCfg(DAL.CFG_PIN_NEOPIXEL);
-        const neonum = control.getConfigValue(DAL.CFG_NUM_NEOPIXELS, 1);
-        if (data && clk) {
-            _onboardStrip = light.createAPA102Strip(data, clk, dsnum);
-            _onboardStrip.setBrightness(96);
-        } else if (neo) {
-            _onboardStrip = light.createNeoPixelStrip(neo, neonum, NeoPixelMode.RGB);
-        } else {
-            _onboardStrip = light.createNeoPixelStrip(undefined, 0);
-        }
-        return _onboardStrip;
-    }
-
     /**
      * Sets the number of LEDS on the default light strip
      */
